START_TEMPLATE/pyproject.toml
[tool.poetry]
name = "start-template"
version = "1.0.0"
description = "Reusable application template with auth, credits, payments, and more"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "shared", from = "backend"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.0"
pydantic = "^2.0.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"
supabase = "^2.0.0"
PyJWT = "^2.8.0"
requests = "^2.31.0"
stripe = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
